[department-of-veterans-affairs/vets-website#227] Renames files for clarity
